implement odic; general refactoring
